requirements.txt
fastapi
uvicorn
pyodbc
mysql-connector-python
